# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Project Management Federation API
  version: 0.1.0
  description: |
    Backend REST, ActivityPub, C2S, S2S inbox, WebFinger, and delivery administration contract.

    Bounded REST list operations return JSON arrays for backward compatibility and expose a uniform
    offset-pagination contract through `Link`, `X-Pagination-Limit`, `X-Pagination-Offset`,
    `X-Pagination-Count`, and `X-Pagination-Has-More` response headers. The `Link` header preserves
    active filters and supplies RFC 8288 `prev` and `next` relations when applicable.
servers:
  - url: http://localhost:8080
...
paths:
...
  /register:
    post:
      tags: [Auth]
      operationId: registerUser
      requestBody:
        $ref: "#/components/requestBodies/RegisterUser"
      responses:
        "201":
          description: User actor created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /login:
    post:
      tags: [Auth]
      operationId: loginUser
      requestBody:
        $ref: "#/components/requestBodies/LoginUser"
      responses:
        "200":
          description: HttpOnly session cookie established.
          headers:
            Set-Cookie:
              description: Hardened Progo browser session cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [user_id, instance_role]
                properties:
                  user_id:
                    type: string
                    format: uuid
                  instance_role:
                    type: string
                    enum: [owner, admin, user]
                  email:
                    type: string
                    format: email
                  email_verified:
                    type: boolean
                  mfa_enrollment_required:
                    type: boolean
        "401":
          $ref: "#/components/responses/Error"
...
  /auth/{provider}/start:
    parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          enum: [google, github]
    get:
      tags: [Auth]
      operationId: startOAuthLogin
      responses:
        "302":
          description: Redirect to the provider authorization URL.
        "404":
          $ref: "#/components/responses/Error"
  /auth/{provider}/callback:
    parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          enum: [google, github]
    get:
      tags: [Auth]
      operationId: completeOAuthProviderCallback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the frontend OAuth callback with a one-time code or safe error code.
  /auth/oauth/exchange:
    post:
      tags: [Auth]
      operationId: exchangeOAuthLoginCode
      requestBody:
        $ref: "#/components/requestBodies/ExchangeOAuthCode"
      responses:
        "200":
          description: HttpOnly session cookie established after consuming a one-time OAuth code.
          content:
            application/json:
              schema:
                type: object
                required: [user_id, instance_role]
                properties:
                  user_id:
                    type: string
                    format: uuid
                  instance_role:
                    type: string
                    enum: [owner, admin, user]
                  email:
                    type: string
                    format: email
                  email_verified:
                    type: boolean
                  mfa_enrollment_required:
                    type: boolean
        "401":
          $ref: "#/components/responses/Error"
...
  /api/v1/me/password:
    patch:
      tags: [Auth]
      operationId: changeCurrentUserPassword
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/ChangePassword"
      responses:
        "204":
          description: Password changed.
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /api/v1/me/logout:
    post:
      tags: [Auth]
      operationId: logoutCurrentUser
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "204":
          description: All issued sessions revoked and browser cookie expired.
        "401":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
...
  /api/v1/admin/users:
    get:
      tags: [Auth]
      operationId: listAdminUsers
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/InstanceRole"
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Local users visible to administrators.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/admin/users/{userID}/role:
    parameters:
      - $ref: "#/components/parameters/UserID"
    patch:
      tags: [Auth]
      operationId: updateAdminUserInstanceRole
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/UpdateInstanceRole"
      responses:
        "200":
          description: User instance role updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: progo.session
      description: HttpOnly SameSite=Strict browser session cookie.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    metricsBearerAuth:
      type: http
      scheme: bearer
      description: Optional in development; required when METRICS_TOKEN is configured.
...
  parameters:
...
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
...
    LoginUser:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [email, password]
            properties:
              email:
                type: string
                format: email
              password:
                type: string
                format: password
              mfa_code:
                type: string
                description: Current TOTP or unused recovery code when MFA is enabled.
...
    ExchangeOAuthCode:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [code]
            properties:
              code:
                type: string
              mfa_code:
                type: string
                description: Current TOTP or unused recovery code when MFA is enabled.
    RegisterUser:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [username, email, password]
            properties:
              username:
                type: string
              email:
                type: string
                format: email
              password:
                type: string
                format: password
    ChangePassword:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [current_password, new_password]
            properties:
              current_password:
                type: string
                format: password
              new_password:
                type: string
                format: password
...
    UpdateInstanceRole:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [instance_role]
            properties:
              instance_role:
                $ref: "#/components/schemas/InstanceRole"
...
  responses:
...
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    User:
      type: object
      required: [id, ap_id, username, email, email_verified, instance_role, handle, name, summary, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        ap_id:
          type: string
          format: uri
        username:
          type: string
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        instance_role:
          $ref: "#/components/schemas/InstanceRole"
        handle:
          type: string
        name:
          type: string
        summary:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    InstanceRole:
      type: string
      enum: [owner, admin, user]
...